turned profile into modal, moved to Profile.js
